Changed the firebase initialization from services to lib
